style: move `RewardFlag` to user_state mod
